app/application/gke-hpa-scaler/requirements.txt
flask==2.2.5
kubernetes==29.0.0
gunicorn==20.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.121.0